services/cloud-configurator/requirements.txt
asgiref==3.2.7
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
click==7.1.2
dataclasses==0.6
dictdiffer==0.8.1
Django==2.2.12
django-mongodb-engine==0.6.0
django-rest-framework-mongoengine==3.4.1
djangotoolbox==1.8.0
djongo==1.3.2
docutils==0.15.2
fastapi==0.58.0
jmespath==0.9.5
jsondiff==1.2.0
lockfile==0.12.2
mongoengine==0.20.0
pymongo==3.9.0
python-daemon==2.2.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
uvicorn==0.11.5
wincertstore==0.2